Heavy Attack (Almost Done) + More Credits Work
Implemented Heavy Attack
 - On Right Click, charge for max of 2 seconds.
 - Does damage based on charge, float since ints are hard to divide
 - Min damage 25, Max damage 100
 - No animation, no limit/cooldown

More Credits Work
 - Fading text in and out working and timed properly
 - Special thanks to Camden County College in
 - Names are a PLACEHOLDER. Proper names are being added soon.
 - Still need to add Professor Morrison thanks and @Anthony Cartelle x3

Implemented Animations
 - "Good" Soul replacing original collectible soul design
 - Heavy Attack fireball sprite sheet added and formatted for animation later

+++ ActualBlueNimbusProject/Assets/Scripts/Enemy.cs
@@ -28,7 +28,7 @@ public class Enemy : MonoBehaviour
     [Header("Player Health Variables")]
     public PlayerCombat playerHealth;
     public int maxHealth = 100;
-    public int currentHealth;
+    public float currentHealth;
 
     public int soulValue;
     void Start()
@@ -45,7 +45,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void EnemyTakeDamage(int damage)
+    public void EnemyTakeDamage(float damage)
     {
         currentHealth -= damage;
